Adding food log page layout fix

+++ lib/data/model/nutrition/food_nutrition_info_model.dart
@@ -1,7 +1,7 @@
 class FoodNutritionInfo {
   String foodId;
   String foodName;
-  int energyKal;
+  double energyKal;
   double proteinG;
   double fatG;
   double carbsG;
